refactor(web): split CommentsSection into MessageItem + shared mention hook
Extract the recursive message row (MessageItem), the pending-upload chips, and
the @mention state machine (useMentions) + dropdown that CommentsSection and
CommentInput each duplicated. CommentsSection drops 488->138 lines and its two
near-identical compose handlers collapse into one. Behavior-preserving.
